updated isolation mode token vault lib for accountNumber=0 swaps

+++ contracts/protocol/interfaces/IDolomiteStructs.sol
@@ -171,7 +171,7 @@ interface IDolomiteStructs {
 
     struct Par {
         bool sign;
-        uint256 value;
+        uint128 value;
     }
 
     struct RiskLimits {
